requirements.txt
langchain
langchain-core
langchain-community
crewai[tools]==0.32.0
# crewai==0.32.0
# crewai-tools==0.3.0
langchain-groq
langchain-openai
langchain-experimental
streamlit

pandas
scipy
matplotlib
python-dotenv